backend_rules/pyproject.toml
[project]
name = "release_candidates_rules"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.121.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
    "uvicorn>=0.38.0",
]
